requirements.txt
openai>=1.66.3
ollama>=0.4.7
rich>=13.9.4
python-dotenv>=1.0.1
python-docx>=1.1.2
markdown>=3.6
htmldocx>=0.0.6